# openapi.json
{
  "components": {
    "schemas": {
...
      "Author": {
        "description": "An author/artist credit.",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/AuthorId"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ],
        "type": "object"
      },
      "AuthorId": {
        "description": "Identifies an author/artist credit.",
        "format": "uuid",
        "type": "string"
      },
...
      "ChapterDto": {
        "properties": {
          "number": {
            "format": "double",
            "type": "number"
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "read": {
            "description": "Whether the requesting user has read this chapter (number ≤ their progress).\n`None` for anonymous requests; `Some(bool)` when a valid `Bearer` token is present\n(frontend §9.2 auth-scoped read-state).",
            "type": [
              "boolean",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "description": "Resolved absolute URL to open the chapter page on the provider.",
            "type": "string"
          }
        },
        "required": [
          "number",
          "url"
        ],
        "type": "object"
      },
...
      "ContentType": {
        "description": "The medium/origin classification of a work.",
        "enum": [
          "manga",
          "manhwa",
          "manhua",
          "webtoon",
          "unknown"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 `application/problem+json` error body shape produced by [`ApiError`]. Declared\npurely for `OpenAPI` documentation — [`ApiError::into_response`] builds the JSON by hand so\nruntime callers never construct this type.",
        "example": {
          "detail": "resource not found",
          "status": 404,
          "title": "not_found",
          "type": "about:blank#not_found"
        },
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
      "SeriesDetail": {
        "properties": {
          "alt_titles": {
            "description": "Alternative titles gathered across providers (frontend §9.2; empty when none).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "anilist_id": {
            "description": "`AniList` media id, if this series is mapped (`sync_mappings`); lets the frontend\nlink out to the canonical `AniList` entry regardless of whether the viewer has synced.",
            "type": [
              "string",
              "null"
            ]
          },
          "authors": {
            "description": "Author/artist credits attached to the series; empty when none.",
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "type": "array"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "cover_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/SeriesId"
          },
          "release_year": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceDto"
            },
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/SeriesStatus"
          },
          "tags": {
            "description": "Genre/tags attached to the series (frontend §9.2; empty when none).",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content_type",
          "status",
          "sources",
          "alt_titles",
          "tags",
          "authors"
        ],
        "type": "object"
      },
      "SeriesId": {
        "description": "Identifies a canonical [`crate::entities::Series`].",
        "format": "uuid",
        "type": "string"
      },
      "SeriesSourceId": {
        "description": "Identifies a [`crate::entities::SeriesSource`].",
        "format": "uuid",
        "type": "string"
      },
      "SeriesStatus": {
        "description": "Publication status of a canonical series.",
        "enum": [
          "ongoing",
          "completed",
          "hiatus",
          "cancelled",
          "unknown"
        ],
        "type": "string"
      },
      "SeriesSummary": {
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "cover_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/SeriesId"
          },
          "source_count": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/SeriesStatus"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content_type",
          "status",
          "source_count"
        ],
        "type": "object"
      },
...
      "Tag": {
        "description": "A genre/tag.",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/TagId"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ],
        "type": "object"
      },
      "TagId": {
        "description": "Identifies a tag.",
        "format": "uuid",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "TankoVault contributors"
    },
    "description": "Axum HTTP edge: public read/write + admin, auth, resolved links, SSE.",
    "license": {
      "name": ""
    },
    "title": "tankovault-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/series": {
      "get": {
        "description": "Filter/sort/paginate the public series list (frontend §9.1). The body remains a plain\n`SeriesSummary[]`; pagination metadata rides on the `X-Total-Count` (rows matching the\nfilter) and `X-Next-Cursor` (next page index, absent on the last page) headers so existing\narray-decoding clients keep working.",
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "content_type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Provider slug the series must have a source on.",
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Tag slugs the series must all carry.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Tag slugs the series must not carry.",
            "in": "query",
            "name": "exclude_tag",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "year_min",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year_max",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "min_chapters",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "`updated | title | chapters | sources | year` (default `updated`).",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Zero-based page index (alias: `cursor`).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SeriesSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Matching series, newest-updated first by default",
            "headers": {
              "X-Next-Cursor": {
                "description": "Next page index; absent on the last page",
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              },
              "X-Total-Count": {
                "description": "Total rows matching the filter",
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Browse the catalogue",
        "tags": [
          "series"
        ]
      }
    },
    "/v1/series/{id}": {
      "get": {
        "operationId": "detail",
        "parameters": [
          {
            "description": "Series id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SeriesId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              }
            },
            "description": "Series detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Series not found"
          }
        },
        "summary": "Get series detail",
        "tags": [
          "series"
        ]
      }
    },
    "/v1/series/{id}/chapters": {
      "get": {
        "description": "Chapter list, newest first. When a valid access token is supplied the per-chapter `read`\nflag is populated from the user's progress (frontend §9.2 auth-scoped read-state);\nanonymous callers get the same list without read-state.",
        "operationId": "chapters",
        "parameters": [
          {
            "description": "Series id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SeriesId"
            }
          },
          {
            "description": "Which source to read chapters from. Defaults to the first source of the series.",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeriesSourceId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Chapters, newest first"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Series (or its source) not found"
          }
        },
        "security": [
          {},
          {
            "bearer_auth": []
          }
        ],
        "summary": "List a source's chapters",
        "tags": [
          "series"
        ]
      }
    },
...
  },
...
}